pyproject.toml
[tool.poetry]
name = "trendbridge"
version = "0.1.0a0"
description = ""
authors = ["Álvaro Vázquez Ortiz <alvvazort@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
atproto = "^0.0.55"
setuptools = "^75.5.0"
python-dotenv = "^1.0.1"
tweepy = "^4.14.0"
python-twitter-v2 = "^0.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
